requirements.txt
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
distlib==0.3.7
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
greenlet==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
playwright==1.39.0
pyee==11.0.1
setuptools==68.2.2
typing_extensions==4.8.0
virtualenv==20.24.6
Werkzeug==3.0.1
